fix: :bug: Engagement dashboard reactivity
Reverted to using useRouterParameter instead of router.getRouteParameters. The reason for this change is that the latter is not reactive and, as a result, it does not update the tab property when switching between tabs. This led to a bug that prevented the engagement dashboard page from properly switching tabs.