feat: Implement users page deactivated tab (#30532)
* feat: :sparkles: Filter users list by active users

Implemented the necessary resources to filter the list of users and to return only those who are active, when the active tab is enabled.

* feat: :sparkles: Implement new roles filter in the admin users page

Implemented a new filter on the users' page that retrieves the roles list and creates a dropdown menu. When any of its options is selected, it will filter the users' list to display only those with the selected roles. Additionally, I made some minor adjustments to the MultiSelectCustom component and introduced some new hooks.

* WIP: Remove filters

* feat: :sparkles: WIP: Implement new actions menu for users page table

* feat: :sparkles: Finish users table menu

Completed the implementation of the users table actions menu by preventing the propagation of the click event that triggered the opening of the contextual bar when the menu was clicked. Also, ensured that the contextual bar only opens when the "Enter" or "Space" keys are used for keyboard navigation and enhanced the options menu in the contextual bar as specified in Figma.

* refactor: :recycle: Remove status badge from contextual bar and reorg

Removed the status from the users page contextual bar (the need to remove this may not be concrete, for now it will be only commented until the final decision is made). Also changed the order of the contextual bar info and changed some minor styles to follow figma specs.

* Typecheck

* Re-add icon back to InfoPanelTitle

* Hide registration status and justify actions menu

* feat: :sparkles: Implement deactivated tab logic and specifications

Implemented the filtering function that will show deactivated users when the deactivated tab has been selected and hid the actions menu options that are unnecessary for deactivated users. Didn't implement the empty state because it has been implemented in the pending tab PR.

* Fix tab types

* Re-add spacing, change components and add save user