feat: :sparkles: Introduce new user.list endpoint to UsersTable
Implemented the new user.list endpoint in the UsersTable.tsx component, now its possible to send a user status to the endpoint and it will return only users with that status. Besides that refactored some hooks and components to better accomodate the new changes. Next will be implementing text search in the new endpoint and if possible the filters.