feat: :sparkles: Implement search by text in new users.list endpoint
Implemented the capability of searching by text in the new users.list endpoint, removed the invited tab from the front-end, refactored the new users.list to also return all users when prompted to and refactored the front end to only use the new endpoint instead of both. Next will solve some smaller issues and implement the filtering if possible.