feat: :sparkles: Implement role filter in new users page
Implemented a filter that fetches all the roles for the current workspace, lists them and allows user to filter the users list by which roles they want to see. Next will be implementing loading logic for the MultiSelectCustom component, the search functionality of the roles filter and checking all options when the all roles option is checked.