feat: :sparkles: Users page table menu and pending tab buttons
Implemented an actions section for the users page table with a menu similar to the one inside of the contextual bar and two new buttons exclusiive to the pending tab, an Activate button for admins to quickly activate users and and Resend Welcome Email for admins to remember users to login for the first time. Also changed some typings to better work with the new code and added the necessary translations.