feat: :sparkles: Introduce reason for joining field in users page contextual bar
Introduced a new field in the contextual bar on the users page named 'Reason for Joining'. This field will display the text that users entered when creating their accounts, but only if the 'Manually Approve New Users' setting is enabled. Additionally, reorganized the contextual bar to enhance clarity and have included the required translation entries in the i18n dictionary.