test: :white_check_mark: Update api tests to follow new listByStatus format
Updated the users.listByStatus tests of the 01-users.js test file to follow the refactor of the listByStatus endpoint. Also removed the 'all' status of the endpoint as it doesn't serve a purpouse anymore and solved a visual glitch in the pending users count that made it blink twice on the first render.