fix: :bug: Blank selected department when selecting out of range
The the Oc current chats departments filter uses "on scroll" pagination. If a user selects a department that isn’t within the first 25 departments listed, the selection will appear blank after refreshing the page. This happens because the `PaginatedSelectFiltered` component cannot find the selected department in the paginated options list due to it not being part of the initially fetched list of departments.