fix: :bug: Version card background image responsivity
The version card background image has been updated to better align with design specifications. On large screens and smaller, the image is fixed to the right side of the container and gradually crops from the right as the container shrinks. On extra-large screens and larger, it remains fixed to the right side with a 20px margin from the container's edge.