BUGFIX: await async function Assets.getTextAsync()
Commit df82bccb86fdab509c9de324ad4adfda93f0ddc1 "chore: Use Assets’ async API (#29311)"
replaced several `Assets.getText()` calls with `await Assets.getTextAsync()`.
One of those getTextAsync calls was not awaited.

That one was especially problematic in that it is in a try/catch block,
however you cannot try/catch a Promise, so we missed handling that error.
You must either try/catch the awaited Promise or chain on a Promise#catch handler.