[Feat] Implement readiness check
Separating liveness and readiness checks into two separate checks.

Liveness we just need something to call to make sure the service is actually alive.

Readiness we need make sure the service is not saturated.  So to do this we are checking the heap size as well as the event loop lag.  These should both be indicators of the system struggling.

To round this out we could could probably add checks for db connection pool, being able to reach db and nats.